+++ beeline/main.js
@@ -93,7 +93,6 @@ app
 .controller('RoutesListController', require('./controllers/RoutesListController.js').default)
 .controller('RouteDetailController', require('./controllers/RouteDetailController.js').default)
 .controller('RouteStopsController', require('./controllers/RouteStopsController.js').default)
-.controller('BookingSpsController', require('./controllers/BookingStopsController.js').default)
 .controller('BookingDatesController', require('./controllers/BookingDatesController.js').default)
 .controller('BookingSummaryController', require('./controllers/BookingSummaryController.js').default)
 .controller('BookingConfirmationController', require('./controllers/BookingConfirmationController.js').default)
@@ -101,10 +100,7 @@ app
 .controller('SettingsController', require('./controllers/SettingsController.js').default)
 .controller('TicketsController', require('./controllers/TicketsController.js').default)
 .controller('TicketDetailController', require('./controllers/TicketDetailController.js').default)
-.controller('TicketDetailController2', require('./controllers/TicketDetailController2.js').default)
-.controller('LiteRouteTrackerController', require('./controllers/LiteRouteTrackerController.js').default)
 .controller('BookingHistoryController', require('./controllers/BookingHistoryController.js').default)
-.controller('LiteSummaryController', require('./controllers/LiteSummaryController.js').default)
 .controller('LiteMoreInfoController', require('./controllers/LiteMoreInfoController.js').default)
 .controller('WelcomeController', require('./controllers/WelcomeController.js').default)
 .controller('KickstarterController', require('./controllers/KickstarterController.js').default)
@@ -114,7 +110,6 @@ app
 .controller('KickstarterRecapController', require('./controllers/KickstarterRecapController.js').default)
 .controller('TabsController', require('./controllers/TabsController.js').default)
 .controller('LiteDetailController', require('./controllers/LiteDetailController.js').default)
-.controller('KickstarterDetailController2', require('./controllers/KickstarterDetailController2.js').default)
 .controller('KickstarterStopsController', require('./controllers/KickstarterStopsController.js').default)
 .directive('searchButton', require('./directives/searchButton.js').default)
 .directive('suggestionViewer', require('./directives/suggestionViewer/suggestionViewer').default)
