--- static/templates/ticket-detail2.html
@@ -1,93 +1,0 @@
-<ion-view class="ticket-detail2" ng-init="activeTab=0">
-  <div class="tabs tabs-top tabs-icon-left tabs-light tabs-striped tabs-color-royal">
-    <a class="tab-item" on-tap="activeTab=0" ng-class="{active: activeTab===0}">
-      <span class="tab-title-large"><i class="ticket-nav-ticket"></i>Ticket</a>
-    </span>
-      <a class="tab-item" on-tap="activeTab=1" ng-class="{active: activeTab===1}">
-        <span class="tab-title-large"><i class="ticket-nav-moreInfo"></i>More Info</span>
-      </a>
-  </div>
-  <ion-content class="has-tabs-top" scroll=false>
-
-    <!-- Ticket Tab -->
-    <div class="trip-info item item-text-wrap" ng-show="activeTab===0">
-      <div class="row">
-        <div class="col col-20">
-          <span class="bus-number" ng-show="route.label">{{route.label}}</span>
-        </div>
-        <div class="col">
-          <h4>Trip Code </h4>
-          <trip-code class="trip-code" code="info.code" ng-if="info.code">
-          </trip-code>
-          <span ng-if="!info.code">Not available</span>
-          <!-- <div class="trip-code"
-            ng-bind="info.trip.id"></div> -->
-        </div>
-        <div class="col">
-          <h4>Vehicle No.</h4>
-          <div ng-if="info.trip.vehicle.vehicleNumber" class="vehicleNo" ng-bind="info.trip.vehicle.vehicleNumber">
-          </div>
-          <div ng-if="!info.trip.vehicle.vehicleNumber">
-            Not provided
-          </div>
-        </div>
-      </div>
-      <div class="row">
-        <div class="route-info">
-          <div class="start">
-            <span class="time" ng-bind="ticket.boardStop.time | date:'h:mm a' : '+0800'"></span><img class="route-icon" src="./img/RoutePairSmallStart.svg" />
-              <div class="routeDetailDesc">
-                {{ticket.boardStop.stop.description}}, {{ticket.boardStop.stop.road}}</div>
-          </div>
-            <div class="end">
-            <span class="time" ng-bind="ticket.alightStop.time | date:'h:mm a' : '+0800'"></span><img class="route-icon" src="./img/RoutePairSmallEnd.svg" />
-              <div class="routeDetailDesc">{{ticket.alightStop.stop.description}}, {{ticket.alightStop.stop.road}}</div>
-          </div>
-        </div>
-      </div>
-    </div>
-
-    <!-- Driver Tab -->
-    <div class="trip-info item item-text-wrap" ng-show="activeTab===1">
-      <div class="row">
-        <div class="col">
-          <h4>Operated By</h4>
-          <span ng-bind="company.name"></span>
-          <a ng-click="showTerms(company.id)"><i class="icon ion-information-circled tnc"></i></a>
-        </div>
-      </div>
-      <div class="row">
-        <div class="col">
-          <h4>Driver</h4>
-          <div ng-bind="info.trip.tripDriver.name  || 'Not provided'"></div>
-        </div>
-        <div class="col">
-          <h4>Passenger</h4>
-          <div ng-bind="user.name"></div>
-        </div>
-      </div>
-    </div>
-
-    <!-- Service cancellation when driver or ops press trip cancellation button, show the following -->
-    <ion-scroll class="service-cancellation" ng-if="info.trip.status == 'cancelled'">
-      <div class="item item-text-wrap">
-        <div class="cancellation-content">
-          <h2>Service is Cancelled</h2>
-          <img src="img/TicketCancelledTrip.svg" />
-          <!-- <i class="icon ion-alert-circled"></i> -->
-          <p>Today's trip has been cancelled. Sorry for the inconvenience caused. </p>
-          <p>
-            Please make alternative transport arrangements.
-            <strong>{{company.name}}</strong> will contact you and refund today's fare.
-          </p>
-          <p>
-            For enquiries, please contact
-            <strong>{{company.email}}</strong>
-            <strong ng-if="company.contactNo">(Tel: {{company.contactNo}})</strong>
-          </p>
-        </div>
-      </div>
-    </ion-scroll>
-    <!-- Service cancellation section ends -->
-  </ion-content>
-</ion-view>
