+++ www/templates/tickets.html
@@ -51,7 +51,7 @@
       <a ng-if="tickets.today.length > 0"
          class="item item-icon-right"
          ng-repeat="ticket in tickets.today"
-         ui-sref="tabs.ticket-detail2({ ticketId: ticket.id })">
+         ui-sref="tabs.ticket-detail({ ticketId: ticket.id })">
         <span class="date">{{ ticket.boardStop.time | date:'d MMM y' : '+0800' }}</span> {{ ticket.boardStop.time | date:'h:mm a' : '+0800' }}
         <i class="icon ion-ios-arrow-right icon-accessory"></i>
       </a>
@@ -60,7 +60,7 @@
       <a ng-if="tickets.soon.length > 0"
          class="item item-icon-right"
          ng-repeat="ticket in tickets.soon"
-         ui-sref="tabs.ticket-detail2({ ticketId: ticket.id })">
+         ui-sref="tabs.ticket-detail({ ticketId: ticket.id })">
         <span class="date">{{ ticket.boardStop.time | date:'d MMM y' : '+0800' }}</span> {{ ticket.boardStop.time | date:'h:mm a' : '+0800' }}
         <i class="icon ion-ios-arrow-right icon-accessory"></i>
       </a>
